services/tester_service/requirements.txt
aiokafka==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
docker==7.1.0 ; python_version >= "3.12" and python_version < "4.0"
ecdsa==0.19.1 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.11 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.8.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-jose==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
pywin32==308 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.39 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.46.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
